Rename `full_proposal` module
